Fix compile error on 64-bit machines.
* This is built for the host system, so we can't use B_PRIdOFF.
* Until POSIX introduces a format constant for off_t, cast the variable
to long long to avoid a warning.

+++ src/tools/data_to_source.cpp
@@ -110,7 +110,7 @@ main(int argc, const char* const* argv)
 
 	// close the braces and write the size variable
 	sprintf(lineBuffer, "};\nlong long %s = %lldLL;\n", sizeVarName,
-		dataSize);
+		(long long)dataSize);
 	write_string(outFD, lineBuffer);
 
 	close(inFD);
